[ty] Refactor go-to definition machinery to allow use from both `ty_python_semantic` and `ty_ide`
