+++ crates/ty_python_semantic/src/types.rs
@@ -181,6 +181,7 @@ mod variance;
 mod visitor;
 
 mod definition;
+pub(crate) mod definition_resolution;
 #[cfg(test)]
 mod property_tests;
 mod subscript;
